fix: изменен тип данных для поля topic_id в миграции
Обновлен тип поля topic_id с Integer на BigInteger в таблице feedback_topics.

+++ src/db/migrations/versions/20260216-0132_dc5dfb23860c_rev_16_add_feedback_tables.py
@@ -22,7 +22,7 @@ def upgrade() -> None:
         "feedback_topics",
         sa.Column("id", sa.BigInteger(), nullable=False, autoincrement=True),
         sa.Column("user_id", sa.BigInteger(), nullable=False),
-        sa.Column("topic_id", sa.Integer(), nullable=False),
+        sa.Column("topic_id", sa.BigInteger(), nullable=False),
         sa.Column(
             "created_at",
             sa.TIMESTAMP(timezone=True),
